make invisible other quest pins while one quest is highlighted
