+++ app/src/main/java/de/westnordost/streetcomplete/map/QuestsMapFragment.kt
@@ -24,6 +24,7 @@ import de.westnordost.streetcomplete.ktx.viewLifecycleScope
 import de.westnordost.streetcomplete.map.components.FocusGeometryMapComponent
 import de.westnordost.streetcomplete.map.components.PinsMapComponent
 import de.westnordost.streetcomplete.map.components.GeometryMarkersMapComponent
+import de.westnordost.streetcomplete.map.components.SelectedPinsMapComponent
 import de.westnordost.streetcomplete.quests.ShowsGeometryMarkers
 import de.westnordost.streetcomplete.util.distanceTo
 import kotlinx.coroutines.Dispatchers
@@ -44,6 +45,7 @@ class QuestsMapFragment : LocationAwareMapFragment(), ShowsGeometryMarkers {
 
     private var geometryMarkersMapComponent: GeometryMarkersMapComponent? = null
     private var pinsMapComponent: PinsMapComponent? = null
+    private var selectedPinsMapComponent: SelectedPinsMapComponent? = null
     private var geometryMapComponent: FocusGeometryMapComponent? = null
     private var questPinsManager: QuestPinsManager? = null
     private var editHistoryPinsManager: EditHistoryPinsManager? = null
@@ -73,16 +75,17 @@ class QuestsMapFragment : LocationAwareMapFragment(), ShowsGeometryMarkers {
         val ctrl = controller ?: return
         ctrl.setPickRadius(1f)
         geometryMarkersMapComponent = GeometryMarkersMapComponent(resources, ctrl)
-        pinsMapComponent = PinsMapComponent(requireContext(), ctrl)
+        pinsMapComponent = PinsMapComponent(ctrl)
+        selectedPinsMapComponent = SelectedPinsMapComponent(requireContext(), ctrl)
         geometryMapComponent = FocusGeometryMapComponent(ctrl)
 
         questPinsManager = QuestPinsManager(ctrl, pinsMapComponent!!, questTypeOrderSource, questTypeRegistry, resources, visibleQuestsSource)
         viewLifecycleOwner.lifecycle.addObserver(questPinsManager!!)
-        questPinsManager!!.isVisible = pinMode == PinMode.QUESTS
+        questPinsManager!!.isActive = pinMode == PinMode.QUESTS
 
         editHistoryPinsManager = EditHistoryPinsManager(pinsMapComponent!!, editHistorySource, resources)
         viewLifecycleOwner.lifecycle.addObserver(editHistoryPinsManager!!)
-        editHistoryPinsManager!!.isVisible = pinMode == PinMode.EDITS
+        editHistoryPinsManager!!.isActive = pinMode == PinMode.EDITS
 
         super.onMapReady()
     }
@@ -139,13 +142,13 @@ class QuestsMapFragment : LocationAwareMapFragment(), ShowsGeometryMarkers {
     /* --------------------------------- Focusing on edit --------------------------------------- */
 
     fun startFocusEdit(edit: Edit, offset: RectF) {
-        pinsMapComponent?.showSelectedPins(edit.icon, listOf(edit.position))
         geometryMapComponent?.beginFocusGeometry(ElementPointGeometry(edit.position), offset)
         geometryMapComponent?.showGeometry(edit.getGeometry())
+        selectedPinsMapComponent?.set(edit.icon, listOf(edit.position))
     }
 
     fun endFocusEdit() {
-        pinsMapComponent?.clearSelectedPins()
+        selectedPinsMapComponent?.clear()
         geometryMapComponent?.endFocusGeometry(returnToPreviousPosition = false)
         geometryMapComponent?.clearGeometry()
     }
@@ -167,17 +170,20 @@ class QuestsMapFragment : LocationAwareMapFragment(), ShowsGeometryMarkers {
     fun startFocusQuest(quest: Quest, offset: RectF) {
         geometryMapComponent?.beginFocusGeometry(quest.geometry, offset)
         geometryMapComponent?.showGeometry(quest.geometry)
-        pinsMapComponent?.showSelectedPins(quest.type.icon, quest.markerLocations)
+        selectedPinsMapComponent?.set(quest.type.icon, quest.markerLocations)
+        // while quest is focussed, we actually don't want to see all the other quest pins (since v38)
+        pinsMapComponent?.isVisible = false
     }
 
     /** Clear focus on current quest but do not return to normal view yet */
     fun clearFocusQuest() {
-        pinsMapComponent?.clearSelectedPins()
+        selectedPinsMapComponent?.clear()
         geometryMapComponent?.clearGeometry()
         geometryMarkersMapComponent?.clear()
     }
 
     fun endFocusQuest() {
+        pinsMapComponent?.isVisible = true
         clearFocusQuest()
         viewLifecycleScope.launch {
             /* small delay to wait for other animations when ending focus on quest to be done first
@@ -189,6 +195,7 @@ class QuestsMapFragment : LocationAwareMapFragment(), ShowsGeometryMarkers {
             geometryMapComponent?.endFocusGeometry()
         }
         centerCurrentPositionIfFollowing()
+
     }
 
     /* -------------------------------  Markers for current quest ------------------------------- */
@@ -212,22 +219,22 @@ class QuestsMapFragment : LocationAwareMapFragment(), ShowsGeometryMarkers {
            may only be activated after the old has been deactivated
          */
         geometryMarkersMapComponent?.clear()
-        pinsMapComponent?.clearSelectedPins()
+        selectedPinsMapComponent?.clear()
         geometryMapComponent?.endFocusGeometry(returnToPreviousPosition = false)
         geometryMapComponent?.clearGeometry()
 
         when (pinMode) {
             PinMode.QUESTS -> {
-                editHistoryPinsManager?.isVisible = false
-                questPinsManager?.isVisible = true
+                editHistoryPinsManager?.isActive = false
+                questPinsManager?.isActive = true
             }
             PinMode.EDITS -> {
-                questPinsManager?.isVisible = false
-                editHistoryPinsManager?.isVisible = true
+                questPinsManager?.isActive = false
+                editHistoryPinsManager?.isActive = true
             }
             else -> {
-                questPinsManager?.isVisible = false
-                editHistoryPinsManager?.isVisible = false
+                questPinsManager?.isActive = false
+                editHistoryPinsManager?.isActive = false
             }
         }
     }
